[FIX][NEW] Reafactor module
