+++ jasper_report/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2017 MultidadosTI (http://www.multidadosti.com.br)
+# @author Aldo Soares <soares_aldo@hotmail.com>
+# License LGPL-3 - See http://www.gnu.org/licenses/lgpl-3.0.html
+
+from . import models
