+++ jasper_report/views/jasper_report_parameters_view.xml
@@ -1,10 +1,11 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <odoo>
-    <record model="ir.ui.view" id="view_jasper_report_parameters">
-        <field name="name">multi_jasper_parameters_form</field>
-        <field name="model">multi_jasper_report_parameters</field>
+
+    <record model="ir.ui.view" id="view_jasper_report_parameters_form">
+        <field name="name">view.jasper.report.parameters.form</field>
+        <field name="model">jasper.report.parameters</field>
         <field name="arch" type="xml">
-            <form string="Multi Jasper Report Parameters">
+            <form string="Jasper Report Parameters">
                 <sheet>
                     <group>
                         <field name="name"/>
@@ -14,4 +15,5 @@
             </form>
         </field>
     </record>
+
 </odoo>
