+++ jasper_report/views/jasper_report_settings.xml
@@ -2,9 +2,9 @@
 <odoo>
     <!-- DBSource -->
 
-        <record model="ir.ui.view" id="view_dbsource_tree">
-            <field name="name">multi_jasper_dbsource_tree</field>
-            <field name="model">multi_jasper_dbsource</field>
+        <record model="ir.ui.view" id="view_jasper_report_db_source_tree">
+            <field name="name">view.jasper.report.db.source.tree</field>
+            <field name="model">jasper.report.db.source</field>
             <field name="type">tree</field>
             <field name="arch" type="xml">
                 <tree string="External DB Sources">
@@ -14,9 +14,9 @@
             </field>
         </record>
 
-        <record model="ir.ui.view" id="view_dbsource_form">
-            <field name="name">multi_jasper_dbsource_form</field>
-            <field name="model">multi_jasper_dbsource</field>
+        <record model="ir.ui.view" id="view_jasper_report_db_source_form">
+            <field name="name">jasper.report.db.source.form</field>
+            <field name="model">jasper.report.db.source</field>
             <field name="arch" type="xml">
                 <form string="External DB Source">
                     <sheet>
@@ -24,7 +24,7 @@
                             <group>
                                 <group>
                                     <field name="name"/>
-                                    <field name="dbname"/>
+                                    <field name="db_name"/>
                                     <field name="host"/>
                                     <field name="port"/>
                                 </group>
@@ -35,31 +35,31 @@
                                     <button name="connection_test" string="Test Connection" type="object" icon="fa-refresh"/>
                                 </group>
                             </group>
-                            <field name="conn_string" invisible="1"/>
+                            <!--<field name="conn_string" invisible="1"/>-->
                         </group>
                     </sheet>
                 </form>
             </field>
         </record>
 
-        <record model="ir.actions.act_window" id="multi_action_dbsource">
+        <record model="ir.actions.act_window" id="action_jasper_report_db_source">
             <field name="name">External Database Sources</field>
-            <field name="res_model">multi_jasper_dbsource</field>
+            <field name="res_model">jasper.report.db.source</field>
             <field name="view_type">form</field>
             <field name="view_mode">tree,form</field>
-            <field name="view_id" ref="view_dbsource_tree"/>
+            <field name="view_id" ref="view_jasper_report_db_source_tree"/>
         </record>
 
-        <menuitem id="multi_jasper_menu" name="Multi Reports" sequence="450"/>
+        <menuitem id="jasper_menu" name="Reports" sequence="450"/>
 
-        <menuitem id="multi_jasper_submenu" name="Multi Reports"
-                  parent="multi_jasper_menu" sequence="1"/>
+        <menuitem id="jasper_submenu" name="Reports"
+                  parent="jasper_menu" sequence="1"/>
 
-        <menuitem id="multi_jasper_config" name="Configurações"
-                  parent="multi_jasper_menu" sequence="2"/>
+        <menuitem id="jasper_config" name="Settings"
+                  parent="jasper_menu" sequence="2"/>
 
-        <menuitem name="Conexões de banco de dados" id="menu_dbsource"
-                  parent="multi_jasper_config"
-                  action="multi_action_dbsource"/>
+        <menuitem name="Database Connections" id="menu_db_source"
+                  parent="jasper_config"
+                  action="action_jasper_report_db_source"/>
 
 </odoo>
