+++ jasper_report/views/jasper_report_view.xml
@@ -1,0 +1,85 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<odoo>
+    <record model="ir.ui.view" id="view_jasper_report_tree">
+        <field name="name">view.jasper.report.tree</field>
+        <field name="model">jasper.report</field>
+        <field name="type">tree</field>
+        <field name="arch" type="xml">
+            <tree string="Jasper Reports">
+                <field name="name"/>
+            </tree>
+        </field>
+    </record>
+
+    <record model="ir.ui.view" id="view_jasper_report_form">
+        <field name="name">view.jasper.report.form</field>
+        <field name="model">jasper.report</field>
+        <field name="arch" type="xml">
+            <form string="Jasper Report">
+                <header>
+                        <button string="Build Report" type="object"
+                                name="get_report" class="oe_highlight"/>
+                        <button string="Download Report" type="object"
+                                name="get_file_report"
+                                attrs="{'invisible': [('pdf_report', '=', False)]}"/>
+                    </header>
+                <sheet>
+                    <group>
+                        <group>
+                            <field name="name"/>
+                            <field name="db_obj"/>
+                            <field name="report_template_filename" invisible="1"/>
+                            <field name="report_template" filename="report_template_filename"/>
+                            <field name="pdf_report_filename" invisible="1"/>
+                            <field name="pdf_report" invisible="1"/>
+                        </group>
+
+                        <!--<group attrs="{'invisible': [('report', '=', False)]}">-->
+                            <!--<p>-->
+                                <!--<b>Parameters</b>-->
+                                <!--<field name="parameters"/>-->
+                                <!--<field name="check_parameters" invisible="1"/>-->
+                                <!--<button string="List Parameters"-->
+                                        <!--type="object"-->
+                                        <!--attrs="{'invisible': [('check_parameters', '=', False)]}"-->
+                                        <!--name="listing_parameters"-->
+                                        <!--class="oe_highlight"/>-->
+                            <!--</p>-->
+                        <!--</group>-->
+                        <!--<group col="4" attrs="{'invisible': [('report', '=', False)]}">-->
+                        <!--<group colspan="2">-->
+                        <!--<button string="Gerar Relatório" type="object" name="get_report" class="oe_highlight"/>-->
+                        <!--<button string="Download Relatório" type="object" name="get_file_report" attrs="{'invisible': [('pdf_report', '=', False)]}"/>-->
+                        <!--</group>-->
+                        <!--</group>-->
+                    </group>
+                    <notebook>
+                            <page string="Parameters">
+                                <field name="parameters"/>
+                                <field name="check_parameters" invisible="1"/>
+                                <button string="List Parameters"
+                                        type="object"
+                                        attrs="{'invisible': [('check_parameters', '=', False)]}"
+                                        name="listing_parameters"
+                                        class="oe_highlight"/>
+
+                            </page>
+                        </notebook>
+                </sheet>
+            </form>
+        </field>
+    </record>
+
+    <record model="ir.actions.act_window" id="action_jasper_report">
+        <field name="name">Jasper Report</field>
+        <field name="res_model">jasper.report</field>
+        <field name="view_type">form</field>
+        <field name="view_mode">tree,form</field>
+        <field name="view_id" ref="view_jasper_report_tree"/>
+    </record>
+
+    <menuitem name="Jasper Report" id="menu_reports"
+              parent="jasper_submenu"
+              action="action_jasper_report"/>
+
+</odoo>
