--- multi_jasper_report/__manifest__.py
@@ -1,18 +1,0 @@
-# -*- coding: utf-8 -*-
-{
-    'name': 'Multi Jasper Report',
-    'version': '10.0.1.0.0',
-    'summary': 'Jasper Client',
-    'category': 'Reports',
-    'author': 'MultidadosTI',
-    'website': 'www.multidadosti.com.br',
-    'license': 'AGPL-3',
-    'depends': [
-        'base',
-    ],
-    'data': [
-        'views/view_settings_jasper_report.xml',
-        'views/view_multi_jasper_report.xml',
-        'views/view_multi_jasper_report_parameters.xml',
-    ],
-}
