--- multi_jasper_report/models/__init__.py
@@ -1,4 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from . import model_settings_jasper_report
-from . import model_multi_jasper_report
