--- multi_jasper_report/views/view_multi_jasper_report.xml
@@ -1,59 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<odoo>
-    <record model="ir.ui.view" id="view_jasper_report_tree">
-        <field name="name">view_jasper_report_tree</field>
-        <field name="model">multi_jasper_report</field>
-        <field name="type">tree</field>
-        <field name="arch" type="xml">
-            <tree string="Multi Jasper Reports">
-                <field name="name"/>
-            </tree>
-        </field>
-    </record>
-
-    <record model="ir.ui.view" id="view_jasper_report_form">
-        <field name="name">multi_jasper_report_form</field>
-        <field name="model">multi_jasper_report</field>
-        <field name="arch" type="xml">
-            <form string="Multi Jasper Report">
-                <sheet>
-                    <group>
-                        <group>
-                            <field name="db_obj"/>
-                            <field name="name"/>
-                            <field name="report"/>
-                            <field name="pdf_report" invisible="1"/>
-                        </group>
-                        <group attrs="{'invisible': [('report', '=', False)]}">
-                            <p>
-                                <b>Parâmetros</b>
-                                <field name="parameters"/>
-                                <field name="check_parameters" invisible="1"/>
-                                <button string="Listar Parâmetros" type="object" attrs="{'invisible': [('check_parameters', '=', False)]}" name="listing_parameters" class="oe_highlight"/>
-                            </p>
-                        </group>
-                        <group col="4" attrs="{'invisible': [('report', '=', False)]}">
-                            <group colspan="2">
-                                <button string="Gerar Relatório" type="object" name="get_report" class="oe_highlight"/>
-                                <button string="Download Relatório" type="object" name="get_file_report" attrs="{'invisible': [('pdf_report', '=', False)]}"/>
-                            </group>
-                        </group>
-                    </group>
-                </sheet>
-            </form>
-        </field>
-    </record>
-
-    <record model="ir.actions.act_window" id="multi_action_jasper_report">
-        <field name="name">Multi Jasper Reports</field>
-        <field name="res_model">multi_jasper_report</field>
-        <field name="view_type">form</field>
-        <field name="view_mode">tree,form</field>
-        <field name="view_id" ref="view_jasper_report_tree"/>
-    </record>
-
-    <menuitem name="Lista de Relatórios" id="menu_reports"
-              parent="multi_jasper_submenu"
-              action="multi_action_jasper_report"/>
-
-</odoo>
